Improved code readability
