+++ V2/vm/src/main/java/com/neverwinterdp/vm/client/shell/SubCommand.java
@@ -1,5 +1,5 @@
 package com.neverwinterdp.vm.client.shell;
-
+//TODO  anthony investigate remove getDescription
 abstract public class SubCommand {
   abstract public void execute(Shell shell, CommandInput cmdInput) throws Exception ;
   abstract public String getDescription();
